excel_to_image/requirements.txt
pandas==2.1.1
Pillow==10.0.1
openpyxl==3.1.2
